commit corrections in hw 2.15
