+++ helpers.js
@@ -1,4 +1,8 @@
-import { nameInputElement, commentInputElement, addButtonElement } from "./main.js"
+import {
+    nameInputElement, commentInputElement, listElement, addButtonElement,
+    userComments, commentsLoading, commentsAdding, addForm, deleteButtonElement
+} from "./main.js";
+
 
 export const safeCode = (string) => {
     return string
@@ -41,4 +45,41 @@ export const delay = (interval = 300) => {
             resolve();
         }, interval);
     });
+}
+
+// вспомогательные функции
+export const commentsLoadActive = () => {
+    commentsLoading.classList.remove("hide");
+}
+
+export const commentsLoadHide = () => {
+    commentsLoading.classList.add("hide");
+}
+
+export const commentsAddActive = () => {
+    commentsAdding.classList.remove("hide");
+}
+
+export const commentsAddHide = () => {
+    commentsAdding.classList.add("hide");
+}
+
+export const addFormActive = () => {
+    addForm.classList.remove("hide");
+    deleteButtonElement.classList.remove("hide");
+}
+
+export const addFormHide = () => {
+    addForm.classList.add("hide");
+    deleteButtonElement.classList.add("hide");
+}       
+
+export const addButtonActive = () => {
+    addButtonElement.disabled = false;
+    addButtonElement.classList.remove("error");
+}
+
+export const addButtonNonActive = () => {
+    addButtonElement.disabled = true;
+    addButtonElement.classList.add("error");
 }
