+++ main.js
@@ -1,28 +1,21 @@
 import { getComments, postComments } from "./api.js";
 import {
-    safeCode,
-    formatDate,
-    clearInput,
-    prepareInput,
-    delay,
-} from "./appfunc.js";
-import {
-    initEventListeners,
-    renderComments,
-    changeLikeButton,
-} from "./mainfunc.js";
-import {
-    commentsLoadActive,
-    commentsLoadHide,
-    commentsAddActive,
-    commentsAddHide,
-    addFormActive,
-    addFormHide,
-    addButtonActive,
-    addButtonNonActive,
-    delButtonActive,
-    delButtonHide,
-} from "./auxilfunc.js";
+  safeCode,
+  formatDate,
+  clearInput,
+  prepareInput,
+  delay,
+  commentsLoadActive,
+  commentsLoadHide,
+  commentsAddActive,
+  commentsAddHide,
+  addFormActive,
+  addFormHide,
+  addButtonActive,
+  addButtonNonActive,
+} from "./helpers.js";
+import { initEventListeners, changeLikeButton } from "./listeners.js";
+import { renderComments } from "./render.js";
 
 export let userComments = [];
 
@@ -38,44 +31,42 @@ export const addForm = document.querySelector(".add-form");
 commentsLoadActive();
 commentsAddHide();
 addFormActive();
-delButtonActive();
 
 function fetchPromiseArr() {
-    getComments()
-        .then((responseData) => {
-            const appComments = responseData.comments.map((comment) => {
-                return {
-                    name: comment.author.name,
-                    date: formatDate(comment.date),
-                    text: comment.text,
-                    likesCounter: comment.likes,
-                    likeButton: false,
-                };
-            });
+  getComments()
+    .then((responseData) => {
+      const appComments = responseData.comments.map((comment) => {
+        return {
+          name: comment.author.name,
+          date: formatDate(comment.date),
+          text: comment.text,
+          likesCounter: comment.likes,
+          likeButton: false,
+        };
+      });
 
-            commentsLoadHide(); // скрываем надпись о загрузке комментариев с сервера
-            commentsAddHide(); // скрываем надпись о добавлении записей на сервер
-            addFormActive(); // активируем форму добавления комментариев
-            delButtonActive(); // активируем кнопку "Удалить последний комментарий"
+      commentsLoadHide(); // скрываем надпись о загрузке комментариев с сервера
+      commentsAddHide(); // скрываем надпись о добавлении записей на сервер
+      addFormActive(); // активируем форму добавления комментариев, активируем кнопку "Удалить последний комментарий"
 
-            userComments = appComments;
-            renderComments();
-            prepareInput();
-        })
-        .catch((error) => {
-            if (error.message === "Сервер недоступен") {
-                alert("Сервер недоступен, попробуйте позже");
-                fetchPromiseArr();
-            } else if (error.message === "Failed to fetch") {
-                alert("Неполадки интернета");
-                addButtonNonActive(); // деактивируем кнопку "Написать",  меняем ее цвет
-                commentsAddHide(); // скрываем надпись о добавлении записей на сервер
-                addFormActive(); // активируем форму добавления комментариев
-            } else {
-                console.log(error);
-            }
-            console.warn(error);
-        });
+      userComments = appComments;
+      renderComments();
+      prepareInput();
+    })
+    .catch((error) => {
+      if (error.message === "Сервер недоступен") {
+        alert("Сервер недоступен, попробуйте позже");
+        fetchPromiseArr();
+      } else if (error.message === "Failed to fetch") {
+        alert("Неполадки интернета");
+        addButtonNonActive(); // деактивируем кнопку "Написать",  меняем ее цвет
+        commentsAddHide(); // скрываем надпись о добавлении записей на сервер
+        addFormActive(); // активируем форму добавления комментариев
+      } else {
+        console.log(error);
+      }
+      console.warn(error);
+    });
 }
 
 fetchPromiseArr();
@@ -87,25 +78,25 @@ addButtonElement.disabled = true;
 addButtonElement.classList.add("error");
 
 nameInputElement.addEventListener("input", (event) => {
-    addButtonActive();
+  addButtonActive();
 });
 
 commentInputElement.addEventListener("input", (event) => {
-    addButtonActive();
+  addButtonActive();
 });
 
 document.addEventListener("keypress", (event) => {
-    if (event.key === "Enter") {
-        addListComment();
-    }
+  if (event.key === "Enter") {
+    addListComment();
+  }
 });
 
 deleteButtonElement.addEventListener("click", (event) => {
-    userComments.splice(userComments.length - 1, 1);
-    renderComments();
-    prepareInput();
-    clearInput();
-    event.stopPropagation();
+  userComments.splice(userComments.length - 1, 1);
+  renderComments();
+  prepareInput();
+  clearInput();
+  event.stopPropagation();
 });
 
 // deleteButtonElement.addEventListener("click", (event) => {
@@ -131,65 +122,62 @@ initEventListeners();
 addButtonElement.addEventListener("click", handlePostClick);
 
 function handlePostClick() {
-    if (
-        nameInputElement.value.trim() === "" ||
-        commentInputElement.value.trim() === ""
-    ) {
-        addButtonNonActive();
-        return;
-    }
+  if (
+    nameInputElement.value.trim() === "" ||
+    commentInputElement.value.trim() === ""
+  ) {
+    addButtonNonActive();
+    return;
+  }
 
-    postComments().then((response) => {
-        if (response.status === 500) throw new Error("Сервер недоступен");
-        if (response.status === 201 || response.status === 200) {
-            fetchPromiseArr();
-            // return response.json();
-        }
-        if (response.status === 400) throw new Error("Неправильный запрос");
+  postComments()
+    .then((response) => {
+      if (response.status === 500) throw new Error("Сервер недоступен");
+      if (response.status === 201 || response.status === 200) {
+        fetchPromiseArr();
+        // return response.json();
+      }
+      if (response.status === 400) throw new Error("Неправильный запрос");
     })
-        // .then(() => {
-        //   return fetchPromiseArr();
-        // })
-        .then(() => {
-            clearInput();
-        })
-        .catch((error) => {
-            if (error.message === "Сервер недоступен") {
-                alert("Сервер недоступен, попробуйте позже");
-                handlePostClick();
+    // .then(() => {
+    //   return fetchPromiseArr();
+    // })
+    .then(() => {
+      clearInput();
+    })
+    .catch((error) => {
+      if (error.message === "Сервер недоступен") {
+        alert("Сервер недоступен, попробуйте позже");
+        handlePostClick();
 
-                addButtonActive();
-                commentsAddHide();
-                addFormActive();
-                delButtonActive();
+        addButtonActive();
+        commentsAddHide();
+        addFormActive();
 
-                renderComments();
-            } else if (error.message === "Неправильный запрос") {
-                alert("Поля 'Имя' и 'Комментарий' должны содержать хотя бы 3 символа");
+        renderComments();
+      } else if (error.message === "Неправильный запрос") {
+        alert("Поля 'Имя' и 'Комментарий' должны содержать хотя бы 3 символа");
 
-                commentsAddHide();
-                addFormActive();
-                delButtonActive();
+        commentsAddHide();
+        addFormActive();
 
-                renderComments();
-                prepareInput();
-            } else if (error.message === "Failed to fetch") {
-                alert("Неполадки интернета");
+        renderComments();
+        prepareInput();
+      } else if (error.message === "Failed to fetch") {
+        alert("Неполадки интернета");
 
-                addButtonActive();
-                commentsAddHide();
-                addFormActive();
-            } else {
-                console.log(error);
-            }
-        });
+        addButtonActive();
+        commentsAddHide();
+      } else {
+        console.log(error);
+      }
+    });
 
-    commentsAddActive();
-    addFormHide();
-    delButtonHide();
+  commentsAddActive();
+  addFormHide();
 
-    renderComments();
-    // clearInput();
+  renderComments();
+  // clearInput();
 }
 
 renderComments();
